Creación de excepciones para controlar endpoint de búsqueda, filtrado y eliminación de datos.

+++ src/main/java/com/arqueobd/arqueobdrf/service/SiteServiceImpl.java
@@ -2,11 +2,13 @@ package com.arqueobd.arqueobdrf.service;
 
 
 import com.arqueobd.arqueobdrf.entity.Site;
+import com.arqueobd.arqueobdrf.error.SiteNotFoundException;
 import com.arqueobd.arqueobdrf.repository.SiteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.Optional;
 
@@ -20,8 +22,12 @@ public class SiteServiceImpl implements SiteService{
 
     //recupera todos los datos de la tabla sites
     @Override
-    public List<Site> findAllSites() {
-        return siteRepository.findAll();
+    public List<Site> findAllSites() throws SiteNotFoundException {
+        List<Site> site = siteRepository.findAll();
+        if(site.isEmpty()){
+            throw new SiteNotFoundException("No se ha encontrado ningún registro en la Base de Datos.");
+        }
+        return site;
     }
 
     @Override
@@ -52,22 +58,49 @@ public class SiteServiceImpl implements SiteService{
     }
 
     @Override
-    public void deleteSite(Long id) {
-        siteRepository.deleteById(id);
+    public void deleteSite(Long id) throws SiteNotFoundException{
+        //verificar si el site con ese id existe
+        Optional<Site> siteOptional = siteRepository.findById(id);
+        if(siteOptional.isPresent()){
+            siteRepository.deleteById(id);
+        }else{
+            throw new SiteNotFoundException("No se ha encontrado ningún registro con el ID "+id+" que eliminar.");
+        }
     }
 
     @Override
-    public Optional<Site> findSiteByWithJPQL(String name) {
-        return siteRepository.findSiteByNameWithJPQL(name);
+    public Optional<Site> findSiteByWithJPQL(String name) throws SiteNotFoundException {
+        Optional<Site> site = siteRepository.findSiteByNameWithJPQL(name);
+        if(!site.isPresent()){
+            throw new SiteNotFoundException("No existe un yacimiento con el nombre de " + name+".");
+        }
+        return site;
     }
 
     @Override
-    public Optional<Site> findByName(String name) {
-        return siteRepository.findByName(name);
+    public Optional<Site> findByName(String name) throws SiteNotFoundException {
+        Optional<Site> site = siteRepository.findByName(name);
+        if(!site.isPresent()){
+            throw new SiteNotFoundException("No existe un yacimiento con el nombre de " + name+".");
+        }
+        return site;
     }
 
     @Override
-    public Optional<Site> findByNameIgnoreCase(String name) {
-        return siteRepository.findByNameIgnoreCase(name);
+    public Optional<Site> findByNameIgnoreCase(String name) throws SiteNotFoundException{
+        Optional<Site> site = siteRepository.findByNameIgnoreCase(name);
+        if(!site.isPresent()){
+            throw new SiteNotFoundException("No existe un yacimiento con el nombre de " + name+".");
+        }
+        return site;
+    }
+
+    @Override
+    public Site findSiteById(Long id) throws SiteNotFoundException {
+        Optional<Site> site = siteRepository.findById(id);
+        if(!site.isPresent()){
+            throw new SiteNotFoundException("No existe un yacimiento con el ID "+id+".");
+        }
+        return site.get();
     }
 }
